+++ src/api/hooks/auth/usePostRegister.ts
@@ -1,8 +1,8 @@
 import { useMutation } from "@tanstack/react-query";
-import { postRegister } from "@api/registerAPI";
 import { RegisterParams } from "@api/types/auth";
 import { useNavigate } from "react-router-dom";
 import routes from "constants/routes";
+import { postRegister } from "@api/authAPI";
 
 const usePostRegister = () => {
   const navigate = useNavigate();
